Fixed backwards stopping

+++ Balancing_Robot_MiniProject/Core/Src/motor.h
@@ -10,12 +10,12 @@
  *  Global Variable and Type Declarations
  *  -------------------------------------------------------------------------------------------------------------
  */
-extern volatile int16_t error_integral;    // Integrated error signal
+extern volatile int error_integral;    // Integrated error signal
 extern volatile uint8_t duty_cycle;    // Output PWM duty cycle
 extern volatile int16_t target_rpm;    // Desired speed target
 extern volatile int16_t motor_speed;   // Measured motor speed
 extern volatile int8_t adc_value;      // ADC measured motor current
-extern volatile int16_t error;         // Speed error signal
+extern volatile int error;         // Speed error signal
 extern volatile uint8_t Kp;            // Proportional gain
 extern volatile uint8_t Ki;            // Integral gain
